+++ devops/aws/launch_task.py
@@ -54,9 +54,9 @@ def container_config(args, task_args, job_name):
         raise ValueError('WANDB_API_KEY not found in .netrc file')
 
     # Get the hugging face key from the cache file
-    hugging_face_key_file = os.path.expanduser("~/.cache/huggingface/token")
-    with open(hugging_face_key_file, 'r') as file:
-        hugging_face_key = file.read().strip()
+    # hugging_face_key_file = os.path.expanduser("~/.cache/huggingface/token")
+    # with open(hugging_face_key_file, 'r') as file:
+    #     hugging_face_key = file.read().strip()
 
     setup_cmds = [
         'git pull',
@@ -103,10 +103,10 @@ def container_config(args, task_args, job_name):
                 'name': 'WANDB_SILENT',
                 'value': 'true'
             },
-            {
-                'name': 'TRANSFORMERS_TOKEN',
-                'value': hugging_face_key
-            },
+            # {
+            #     'name': 'TRANSFORMERS_TOKEN',
+            #     'value': hugging_face_key
+            # },
             {
                 'name': 'COLOR_LOGGING',
                 'value': 'false'
